feat: windows support (#24)
* fix(build): updated just to only set powershell as the 'windows-shell'

* ci: added windows ci

* refactor: replaced jackc_next_source_file with a jackc_dir_iterator to minimize static state

added jackc_has_extension to file utils
added jackc_join_path to file utils
added Allocator as an argument to more file util functions

* refactor: diagnostic-engine test runners no longer depend on linux api

fixed vm_parser_tests not being a part of unit-tests
replaced PLATFORM in CMAKE in favor of CMAKE_SYSTEM_NAME
removed stdlib symlink in e2e tests as windows was unable to correctly interpret it
fixed invalid remove-item command in just clean recipe for windows

* refactor: removed __COUNTER__ from tests and "inlined" tests configuration arrays into the REGISTER_TEST

* build: added explicit --compiler flag to just build to specify a compiler to use

* refactor: moved PATH_MAX to a separate header

PATH_MAX is now #undef before usage to avoid redeclaration

+++ jackc/core/allocators/allocators.h
@@ -7,7 +7,7 @@
 #define ALIGN_UP(round, to_multiple_of) (((round) + ((to_multiple_of) - 1)) & ~((to_multiple_of) - 1))
 
 #ifdef __rars__
-#   define ALIGNMENT 4
+#   define ALIGNMENT (unsigned int)4
 #else
 #   define ALIGNMENT alignof(max_align_t)
 #endif
